refactor: signin, signup, and password reset
