+++ src/components/auth/index.ts
@@ -1,0 +1,10 @@
+export interface AuthFormProps {
+  setForm: (from: FormKind) => void;
+}
+
+export enum AuthFormKind {
+  SignIn = "Sign In",
+  SignUp = "Sign Up",
+  ResetPassword = "Reset Password",
+  AuthFormKind = "AuthFormKind",
+}
